Add experimental support for images (github issue #3) Fix for #links (github issue #5)
If we click on a #link, we simply move to that link and don't reload the
page again.

Experimental support for images now clones all images and after
tranquility processing is complete, adds back images at appropriate
locations if they are not already included (and are not ad related
images).